+++ CHANGELOG.md
@@ -7,5 +7,6 @@ All Notable changes to `drupal/opening-hours` module.
 ### Added
 * DMOH-19: Added service configuration form.
 * DMOH-20: Added the opening hours field type.
+* DMOH-21: Added the opening hours field widget.
 
 [Unreleased]: https://github.com/StadGent/drupal_module_opening-hours/compare/master...develop